LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/centering-expected.txt


FAIL horizontal-tb: tall viewport assert_equals: top expected 45 but got 8
FAIL horizontal-tb: wide viewport assert_equals: top expected 15 but got 8
FAIL horizontal-tb: square viewport assert_equals: top expected 45 but got 8
FAIL horizontal-tb: dialog and viewport match assert_equals: top expected 0 but got 8
FAIL vertical-rl: tall viewport assert_equals: left expected 10 but got 12
FAIL vertical-lr: tall viewport assert_equals: left expected 10 but got 8
FAIL vertical-rl (dialog horizontal-tb): tall viewport assert_equals: top expected 45 but got 8
FAIL vertical-lr (dialog horizontal-tb): tall viewport assert_equals: top expected 45 but got 8
FAIL horizontal-tb (container vertical-rl): tall viewport assert_equals: left expected 10 but got 12
FAIL vertical-rl (container horizontal-tb): tall viewport assert_equals: top expected 45 but got 8
FAIL horizontal-tb (container vertical-rl) (dialog horizontal-tb): tall viewport assert_equals: top expected 45 but got 8
FAIL vertical-rl (container horizontal-tb) (dialog vertical-rl): tall viewport assert_equals: left expected 10 but got 12
